Use AddressBadge everywhere
